requirements.txt
psutil
tkinter
mysql-connector-python